requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
discord==1.7.3
discord.py==1.7.3
idna==3.3
multidict==6.0.2
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==0.19.2
requests==2.27.1
toml==0.10.2
typing-extensions==4.1.1
urllib3==1.26.8
yarl==1.7.2
youtube-dl==2021.12.17
